cleaning the old function in the sensitivity study directory
