+++ SensitivityStudy/ConstantsSnippet.h
@@ -1,0 +1,20 @@
+#ifndef _SSConstants_H_
+#define _SSConstants_H_
+
+#include "DC_sb_LL_Header.h"
+
+//#define NSB 45
+//#define NSB 37
+#define NSB 59
+//#define NSB 84
+
+//define lumi in pb-1
+//2015 data 13TeV
+//#define LUMI 2262.946
+//2016 ICHEP Exp
+//#define LUMI 8000
+//2016 End of year Exp
+#define LUMI 35000
+//#define LUMI 12918.140703927
+
+#endif
