--- SensitivityStudy/SSPlots.cc
@@ -1,38 +1,0 @@
-#include <vector>
-#include <iostream>
-#include <string>
-#include <cstring>
-#include <set>
-
-#include "TFile.h"
-#include "TList.h"
-#include "TString.h"
-#include "TH1D.h"
-#include "TCanvas.h"
-#include "TPad.h"
-#include "TLegend.h"
-#include "TLatex.h"
-#include "TStyle.h"
-
-#include "SSPlots.h"
-
-int main(int argc, char* argv[])
-{
-
-  if (argc < 1)
-  {
-    std::cerr <<"Please give at least 1 argument " << "TargetDirName" << std::endl;
-    std::cerr <<" Valid configurations are " << std::endl;
-    std::cerr <<" ./SSPlots SSPlots20160504" << std::endl;
-    return -1;
-  }
-
-  std::string DirName = argv[1];
-
-  SSPlots mySSPlots;
-  mySSPlots.Initialization(DirName);
-  mySSPlots.SSPlotsLoop();
-  
-  return 0;
-}
-
